allow right click on table
Previously, a user had to first left-click on a table
to select it, and then they were able to right-click
on the table to activate the custom context menu
for the table. This PR allows a user to right click
directly on a table to access the custom context
menu without first left clicking the table.